reorganizing vectorizers
