--- Redis.OM.Vectorizers.OpenAI/Redis.OM.Vectorizers.OpenAI.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <RootNamespace>Redis.OM.Vectorizers.OpenAI</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\src\Redis.OM.Vectorizers.Core\Redis.OM.Vectorizers.Core.csproj" />
-      <ProjectReference Include="..\src\Redis.OM\Redis.OM.csproj" />
-    </ItemGroup>
-
-</Project>
