+++ Redis.OM.sln
@@ -10,15 +10,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Redis.OM.POC", "src\Redis.O
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Redis.OM.Unit.Tests", "test\Redis.OM.Unit.Tests\Redis.OM.Unit.Tests.csproj", "{570BF479-BCF4-4D1B-A702-2234CA0A3E7D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Redis.OM.Vectorizers.HuggingFace", "src\Redis.OM.Vectorizers.HuggingFace\Redis.OM.Vectorizers.HuggingFace.csproj", "{329F600A-3AF7-456A-8652-A79A39804EE5}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Redis.OM.Vectorizers.OpenAI", "Redis.OM.Vectorizers.OpenAI\Redis.OM.Vectorizers.OpenAI.csproj", "{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Redis.OM.Vectorizers.Core", "src\Redis.OM.Vectorizers.Core\Redis.OM.Vectorizers.Core.csproj", "{4B9F4623-3126-48B7-B690-F28F702A4717}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Redis.OM.Vectorizers", "src\Redis.OM.Vectorizers\Redis.OM.Vectorizers.csproj", "{4B9F4623-3126-48B7-B690-F28F702A4717}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Vectorizers", "Vectorizers", "{452DC80B-8195-44E8-A376-C246619492A8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Redis.OM.Vectorizers.AzureOpenAI", "src\Redis.OM.Vectorizers.AzureOpenAI\Redis.OM.Vectorizers.AzureOpenAI.csproj", "{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -66,30 +60,6 @@ Global
 		{570BF479-BCF4-4D1B-A702-2234CA0A3E7D}.Release|x64.Build.0 = Release|Any CPU
 		{570BF479-BCF4-4D1B-A702-2234CA0A3E7D}.Release|x86.ActiveCfg = Release|Any CPU
 		{570BF479-BCF4-4D1B-A702-2234CA0A3E7D}.Release|x86.Build.0 = Release|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Debug|x64.Build.0 = Debug|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Debug|x86.Build.0 = Debug|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Release|x64.ActiveCfg = Release|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Release|x64.Build.0 = Release|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Release|x86.ActiveCfg = Release|Any CPU
-		{329F600A-3AF7-456A-8652-A79A39804EE5}.Release|x86.Build.0 = Release|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Debug|x64.Build.0 = Debug|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Debug|x86.Build.0 = Debug|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Release|x64.ActiveCfg = Release|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Release|x64.Build.0 = Release|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Release|x86.ActiveCfg = Release|Any CPU
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19}.Release|x86.Build.0 = Release|Any CPU
 		{4B9F4623-3126-48B7-B690-F28F702A4717}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4B9F4623-3126-48B7-B690-F28F702A4717}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4B9F4623-3126-48B7-B690-F28F702A4717}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -102,18 +72,6 @@ Global
 		{4B9F4623-3126-48B7-B690-F28F702A4717}.Release|x64.Build.0 = Release|Any CPU
 		{4B9F4623-3126-48B7-B690-F28F702A4717}.Release|x86.ActiveCfg = Release|Any CPU
 		{4B9F4623-3126-48B7-B690-F28F702A4717}.Release|x86.Build.0 = Release|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Debug|x64.Build.0 = Debug|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Debug|x86.Build.0 = Debug|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Release|x64.ActiveCfg = Release|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Release|x64.Build.0 = Release|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Release|x86.ActiveCfg = Release|Any CPU
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -123,9 +81,6 @@ Global
 		{E3A31119-E4F1-4793-B5C2-ED2D51502B01} = {8D9ECCFF-E022-4B68-BB43-228CEA248DEC}
 		{452DC80B-8195-44E8-A376-C246619492A8} = {8D9ECCFF-E022-4B68-BB43-228CEA248DEC}
 		{4B9F4623-3126-48B7-B690-F28F702A4717} = {452DC80B-8195-44E8-A376-C246619492A8}
-		{329F600A-3AF7-456A-8652-A79A39804EE5} = {452DC80B-8195-44E8-A376-C246619492A8}
-		{D9797D53-E8E2-4D60-9B07-2CA087E6CD19} = {452DC80B-8195-44E8-A376-C246619492A8}
-		{E64C4D96-9AA1-4C4A-895E-E7A232F29D0D} = {452DC80B-8195-44E8-A376-C246619492A8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {E5752441-184B-4F17-BAD0-93823AC68607}
