--- src/Redis.OM.Vectorizers.AzureOpenAI/Redis.OM.Vectorizers.AzureOpenAI.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Redis.OM.Vectorizers.Core\Redis.OM.Vectorizers.Core.csproj" />
-      <ProjectReference Include="..\Redis.OM\Redis.OM.csproj" />
-    </ItemGroup>
-
-</Project>
