+++ src/Redis.OM.Vectorizers/AzureOpenAISentenceVectorizer.cs
@@ -4,7 +4,7 @@ using System.Text.Json;
 using Redis.OM.Contracts;
 using Redis.OM.Modeling;
 
-namespace Redis.OM.Vectorizers.AzureOpenAI;
+namespace Redis.OM.Vectorizers;
 
 public class AzureOpenAISentenceVectorizer : IVectorizer<string>
 {
