+++ src/Redis.OM.Vectorizers/Configuration.cs
