+++ test/Redis.OM.Unit.Tests/Address.cs
@@ -5,7 +5,6 @@ using System.Runtime.InteropServices;
 using System.Text;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
-using Redis.OM;
 using Redis.OM.Modeling;
 
 namespace Redis.OM.Unit.Tests
