+++ test/Redis.OM.Unit.Tests/Redis.OM.Unit.Tests.csproj
@@ -14,10 +14,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Redis.OM.Vectorizers.OpenAI\Redis.OM.Vectorizers.OpenAI.csproj" />
-    <ProjectReference Include="..\..\src\Redis.OM.Vectorizers.AzureOpenAI\Redis.OM.Vectorizers.AzureOpenAI.csproj" />
-    <ProjectReference Include="..\..\src\Redis.OM.Vectorizers.HuggingFace\Redis.OM.Vectorizers.HuggingFace.csproj" />
     <ProjectReference Include="..\..\src\Redis.OM.POC\Redis.OM.POC.csproj" />
+    <ProjectReference Include="..\..\src\Redis.OM.Vectorizers\Redis.OM.Vectorizers.csproj" />
   </ItemGroup>
 
 </Project>
